+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Z-projection module: check that output filenames are unique (i.e. to avoid two input files with same output filename).
 * Segmentation module: check that output filenames are unique (i.e. to avoid two input files with same output filename).
 * Cell tracking module: check that output filenames are unique (i.e. to avoid two input files with same output filename).
+* Graph filtering module: check that output filenames are unique (i.e. to avoid two input files with same output filename).
 
 ### Changed